+++ drift_dev/lib/src/cli/commands/schema/dump.dart
@@ -1,11 +1,15 @@
 import 'dart:convert';
 import 'dart:io';
+import 'dart:typed_data';
 
 import 'package:args/command_runner.dart';
+import 'package:collection/collection.dart';
 import 'package:path/path.dart';
+import 'package:sqlite3/sqlite3.dart';
 
 import '../../../analysis/results/results.dart';
 import '../../../services/schema/schema_files.dart';
+import '../../../services/schema/sqlite_to_drift.dart';
 import '../../cli.dart';
 
 class DumpSchemaCommand extends Command {
@@ -35,11 +39,68 @@ class DumpSchemaCommand extends Command {
       usageException('Expected input and output files');
     }
 
+    final absolute = File(rest[0]).absolute;
+    final _AnalyzedDatabase result;
+    if (await absolute.isSqlite3File) {
+      result = await _readElementsFromDatabase(absolute);
+    } else {
+      result = await _readElementsFromSource(absolute);
+    }
+
+    final writer =
+        SchemaWriter(result.elements, options: cli.project.moorOptions);
+
+    var target = rest[1];
+    // This command is most commonly used to write into
+    // `<dir>/drift_schema_vx.json`. When we get a directory as a second arg,
+    // try to infer the file name.
+    if (await FileSystemEntity.isDirectory(target) ||
+        !target.endsWith('.json')) {
+      final version = result.schemaVersion;
+
+      if (version == null) {
+        // Couldn't read schema from database, so fail.
+        usageException(
+          'Target is a directory and the schema version could not be read from '
+          'the database class. Please use a full filename (e.g. '
+          '`$target/drift_schema_v3.json`)',
+        );
+      }
+
+      target = join(target, 'drift_schema_v$version.json');
+    }
+
+    final file = File(target).absolute;
+    final parent = file.parent;
+    if (!await parent.exists()) {
+      await parent.create(recursive: true);
+    }
+
+    await file.writeAsString(json.encode(writer.createSchemaJson()));
+    print('Wrote to $target');
+  }
+
+  /// Reads available drift elements from an existing sqlite database file.
+  Future<_AnalyzedDatabase> _readElementsFromDatabase(File database) async {
+    final opened = sqlite3.open(database.path);
+
+    try {
+      final elements = await extractDriftElementsFromDatabase(opened);
+      final userVersion = opened.select('pragma user_version').single[0] as int;
+
+      return _AnalyzedDatabase(elements, userVersion);
+    } finally {
+      opened.dispose();
+    }
+  }
+
+  /// Extracts available drift elements from a [dart] source file defining a
+  /// drift database class.
+  Future<_AnalyzedDatabase> _readElementsFromSource(File dart) async {
     final driver = await cli.createMoorDriver();
 
-    final absolute = File(rest[0]).absolute.path;
     final input =
-        await driver.driver.fullyAnalyze(driver.uriFromPath(absolute));
+        await driver.driver.fullyAnalyze(driver.uriFromPath(dart.path));
 
     if (!input.isFullyAnalyzed) {
       cli.exit('Unexpected error: The input file could not be analyzed');
@@ -56,36 +117,37 @@ class DumpSchemaCommand extends Command {
     final databaseElement = databases.single;
     final db = result.resolvedDatabases[databaseElement.id]!;
 
-    final writer =
-        SchemaWriter(db.availableElements, options: cli.project.moorOptions);
+    return _AnalyzedDatabase(
+        db.availableElements, databaseElement.schemaVersion);
+  }
+}
 
-    var target = rest[1];
-    // This command is most commonly used to write into
-    // `<dir>/drift_schema_vx.json`. When we get a directory as a second arg,
-    // try to infer the file name.
-    if (await FileSystemEntity.isDirectory(target) ||
-        !target.endsWith('.json')) {
-      final version = databaseElement.schemaVersion;
+extension on File {
+  static final _headerStart = ascii.encode('SQLite format 3\u0000');
 
-      if (version == null) {
-        // Couldn't read schema from database, so fail.
-        usageException(
-          'Target is a directory and the schema version could not be read from '
-          'the database class. Please use a full filename (e.g. '
-          '`$target/drift_schema_v3.json`)',
-        );
+  /// Checks whether the file is probably a sqlite3 database file by looking at
+  /// the initial bytes of the expected header.
+  Future<bool> get isSqlite3File async {
+    final opened = await open();
+
+    try {
+      final bytes = Uint8List(_headerStart.length);
+      final bytesRead = await opened.readInto(bytes);
+
+      if (bytesRead < bytes.length) {
+        return false;
       }
 
-      target = join(target, 'drift_schema_v$version.json');
+      return const ListEquality<int>().equals(_headerStart, bytes);
+    } finally {
+      await opened.close();
     }
-
-    final file = File(target);
-    final parent = file.parent;
-    if (!await parent.exists()) {
-      await parent.create(recursive: true);
-    }
-
-    await File(target).writeAsString(json.encode(writer.createSchemaJson()));
-    print('Wrote to $target');
   }
+}
+
+class _AnalyzedDatabase {
+  final List<DriftElement> elements;
+  final int? schemaVersion;
+
+  _AnalyzedDatabase(this.elements, this.schemaVersion);
 }